Fix bug #1227984 - now autosave is based on realtime again, however if the game is paused, no autosaving is triggered, but the time is reset to be retriggered 30 seconds later. The fix includes a fix of gametime variabels from int32_t to uint32_t - this is because SDL_GetTicks is uint32_t as well. some follow parameters are still int32_t leading to some warnings during compilation - however functionality is still 100% the same as the actual value of gametime was positive before in every way anyways - i will fix these warnings later, but did not want to mix them up in this commit.

+++ src/economy/request.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2002-2004, 2006-2011 by the Widelands Development Team
+ * Copyright (C) 2002-2004, 2006-2011, 2015 by the Widelands Development Team
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -310,7 +310,7 @@ int32_t Request::get_priority (int32_t cost) const
 		MAX_IDLE_PRIORITY
 		+
 		std::max
-			(1,
+			(uint32_t(1),
 			 ((m_economy->owner().egbase().get_gametime() -
 			   (is_construction_site ?
 			    get_required_time() : get_last_request_time()))
